changed file structure and namespace name to fix naming conflicts between the function borrow and the namespace borrow. Also to make the namespace name make more sense
