+++ .gitignore
@@ -2,4 +2,5 @@
 *.out
 *.cpp
 !examples/*
+!tests/*
 build
